Fix issue with CPU usage
The port from the old crossbeam::queue::MsQueue to
crossbeam::queue::SegQueue naively had the `run_thread` function poll
the SegQueue in a hot spinning loop. The MsQueue's behaviour would be
that the `pop()` call would block until there is an item in the queue.

The port was due to a security vulnerability in crossbeam that had
removed MsQueue by the time the vulnerability was patched in the library.

To resolve this issue, a custom `BlockingQueue` was implemented in this
library that provides the blocking capabilities via crossbeam channels.

This queue replaces the SegQueue and the MsQueue in the PoolInner struct.